requirements.txt
mistralai
telethon
python-dotenv